LayoutTests/touchadjustment/iframe-boundary.html

<!DOCTYPE html>
<html>
<head>
    <title>Touch Adjustment : Test touch adjustment at an iframe boundary - bug 94936</title>
    <script src="../fast/js/resources/js-test-pre.js"></script>
    <script src="resources/touchadjustment.js"></script>
    <style>
        #navbar {
            background: #eee;
            height: 100%;
            left: 0;
            padding: 0;
            position: absolute;
            margin: 0;
            top: 0;
            width: 100px;
        }

        #content-page {
            height: 100%;
            left: 100px;
            position: absolute;
            right: 0;
            top: 0;
        }

        iframe {
            border: none;
            width: 100%;
            height: 100%;
            margin: 0;
            padding: 0;
        }
    </style>
</head>
<body id="main" onload="runTests()">
<div id="navbar"  style="-webkit-transform: translateX(0px);">
    <iframe id="nav-frame" src="resources/inner-navigation-frame.html"></iframe>
</div>
<div id="content-page">
    <iframe id="content-frame" src="resources/inner-content-page.html"></iframe>
</div>

<p id='description'></p>
<div id='console'></div>

<script> 
    var itemHeight = 32;
    var borderOffset = 95;
    var checkboxOffset = 15;
    var linkOffset = -15;
    var checkboxOffsetIndirect = 5;
    var linkOffsetIndirect = -5;
    var touchRadius = 30;

    function testRoundTouch(x, y, radius)
    {
        var x = x - radius;
        var y = y - radius;
        var width = radius * 2;
        var height = radius * 2;
        var adjustedNode = internals.touchNodeAdjustedToBestClickableNode(x, y, width, height, document);
        if (adjustedNode.nodeType == 3) // TEXT node
            adjustedNode = adjustedNode.parentNode;
        return adjustedNode;
    }
    function testDirectTouches()
    {
        debug('Test direct touches.');
        for (var i = 0; i < 4; i++) {
            adjustedNode = testRoundTouch(borderOffset + checkboxOffset, (i + 0.5) * itemHeight, touchRadius);
            shouldBeEqualToString('adjustedNode.id', 'checkbox-' + (i + 1));
            adjustedNode = testRoundTouch(borderOffset + linkOffset, (i + 0.5) * itemHeight, touchRadius);
            shouldBeEqualToString('adjustedNode.id', 'button-' + (i + 1));
        }
    }
    function testIndirectTouches() {
         debug('Test indirect touches.');
         for (var i = 0; i < 4; i++) {
            adjustedNode = testRoundTouch(borderOffset + checkboxOffsetIndirect, (i + 0.5) * itemHeight, touchRadius);
            shouldBeEqualToString('adjustedNode.id', 'checkbox-' + (i + 1));
            adjustedNode = testRoundTouch(borderOffset + linkOffsetIndirect, (i + 0.5) * itemHeight, touchRadius);
            shouldBeEqualToString('adjustedNode.id', 'button-' + (i + 1));
        }
    }

    function runTests()
    {
        if (window.testRunner && window.internals && internals.touchNodeAdjustedToBestClickableNode) {
            description('Test touch-adjustment at the boundary of an <iframe>.  Ensure that touch adjustment propagates into children of an <iframe>.  Test expected to contain failures until bug 95204 is fixed.');
            testDirectTouches();
            testIndirectTouches();
            finishJSTest();
        }
    }
    jsTestIsAsync = true;
</script>
<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
