LayoutTests/touchadjustment/iframe-boundary-expected.txt

 
Test touch-adjustment at the boundary of an 


Test direct touches.
PASS adjustedNode.id is "checkbox-1"
PASS adjustedNode.id is "button-1"
PASS adjustedNode.id is "checkbox-2"
PASS adjustedNode.id is "button-2"
PASS adjustedNode.id is "checkbox-3"
PASS adjustedNode.id is "button-3"
PASS adjustedNode.id is "checkbox-4"
PASS adjustedNode.id is "button-4"
Test indirect touches.
PASS adjustedNode.id is "checkbox-1"
PASS adjustedNode.id is "button-1"
PASS adjustedNode.id is "checkbox-2"
PASS adjustedNode.id is "button-2"
PASS adjustedNode.id is "checkbox-3"
PASS adjustedNode.id is "button-3"
PASS adjustedNode.id is "checkbox-4"
PASS adjustedNode.id is "button-4"
PASS successfullyParsed is true

TEST COMPLETE
